requirements.txt
aliyun-python-sdk-core-v3==2.9.4
aliyun-python-sdk-dysmsapi==1.0.0
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.1.2
djangorestframework==3.9.0
idna==2.7
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
python-memcached==1.59
pytz==2018.5
qiniu==7.2.2
requests==2.20.0
six==1.11.0
urllib3==1.24
